+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cli: Allowed the `migrate referral-code` subcommand to accept multiple addresses and allow the use of user account addresses or owner account addresses.
 - cli: Ensured all commands respect the `--priority-lamports` option.
 - cli: The `inspect glv` command now supports querying all existing valid GLV accounts.
+- cli: Included GLV-related addresses when extending LUTs with market kind.
 
 ## [0.4.0] - 2025-03-08